Crate aidoku_helpers

Utilities

Utilities to make coding Aidoku sources less miserable. It is available by enabling the helpers feature in the main crate

aidoku = { git = "https://github.com/Aidoku/aidoku-rs", features = ["helpers"] }

After enabling the helpers feature, this crate is re-exported under the aidoku::helpers namespace.

Modules

cfemail

Handles Cloudflare email protection.

node

substring

Substring utility functions.

uri

Module for encoding URIs.