Trait aidoku_helpers::substring::Substring

pub trait Substring {
    // Required methods
    fn substring_before<P: Pattern>(&self, pat: P) -> Option<&str>;
    fn substring_before_last<P: Pattern>(&self, pat: P) -> Option<&str>
       where for<'a> P::Searcher<'a>: ReverseSearcher<'a>;
    fn substring_after<P: Pattern>(&self, pat: P) -> Option<&str>;
    fn substring_after_last<'a, P>(&'a self, pat: P) -> Option<&'a str>
       where P: Pattern,
             <P as Pattern>::Searcher<'a>: ReverseSearcher<'a>;
}

Required Methods

fn substring_before<P: Pattern>(&self, pat: P) -> Option<&str>

Returns a substring before the first occurence of pattern.

fn substring_before_last<P: Pattern>(&self, pat: P) -> Option<&str>
where for<'a> P::Searcher<'a>: ReverseSearcher<'a>,

Returns a substring before the last occurence of pattern.

fn substring_after<P: Pattern>(&self, pat: P) -> Option<&str>

Returns a substring after the first occurence of pattern.

fn substring_after_last<'a, P>(&'a self, pat: P) -> Option<&'a str>
where P: Pattern, <P as Pattern>::Searcher<'a>: ReverseSearcher<'a>,

Returns a substring after the last occurence of pattern.

Object Safety

This trait is not object safe.