Trait aidoku_helpers::substring::Substring

pub trait Substring {
    fn substring_before<'a, P: Pattern<'a>>(&'a self, pat: P) -> Option<&'a str>;
    fn substring_before_last<'a, P>(&'a self, pat: P) -> Option<&'a str>
    where
        P: Pattern<'a>,
        <P as Pattern<'a>>::Searcher: ReverseSearcher<'a>;
    fn substring_after<'a, P: Pattern<'a>>(&'a self, pat: P) -> Option<&'a str>;
    fn substring_after_last<'a, P>(&'a self, pat: P) -> Option<&'a str>
    where
        P: Pattern<'a>,
        <P as Pattern<'a>>::Searcher: ReverseSearcher<'a>;
}

Required Methods

fn substring_before<'a, P: Pattern<'a>>(&'a self, pat: P) -> Option<&'a str>

Returns a substring before the first occurence of pattern.

fn substring_before_last<'a, P>(&'a self, pat: P) -> Option<&'a str> where
P: Pattern<'a>,
<P as Pattern<'a>>::Searcher: ReverseSearcher<'a>,

Returns a substring before the last occurence of pattern.

fn substring_after<'a, P: Pattern<'a>>(&'a self, pat: P) -> Option<&'a str>

Returns a substring after the first occurence of pattern.

fn substring_after_last<'a, P>(&'a self, pat: P) -> Option<&'a str> where
P: Pattern<'a>,
<P as Pattern<'a>>::Searcher: ReverseSearcher<'a>,

Returns a substring after the last occurence of pattern.